+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - es6 `be.map`, Map object assertion
 - es6 `be.mapIterator`, Map Iterator object assertion
 - es6 `be.setIterator`, Set Iterator object assertion
+- es6 `be.int8Array`, Int8Array object assertion
 
 ## [1.4.1] - 2017-07-10